// original/src/modules/JobManagement/JobList/JobList.types.ts
export interface Job {
  jobId: string;
  orderId: string;
  ascId: string;
  source: string;
  pickupType: string;
  paymentType?: string;
  internalReferenceNumber?: string;

  customer: {
    firstName: string;
    lastName: string;
    customerType?: string;
    companyName?: string;
    dealershipName?: string;
    phoneNumber?: string;
    mobileNumber?: string;
  };

  createdAt: string;
  updatedAt: string;

  ascName?: string;
  assigneeName: string | null;
  jobStatus: string;
  isOnHold?: boolean;

  customerWish?: string;

  diagnosticInfo?: {
    actionType?: string;
    materialCost?: number;
    materialsJobType: string[];
  };

  asset?: {
    bareToolNumber?: string;
    serialNumber?: string;
    toolModelName?: string;
    category?: string;
    categoryId?: string;
  };

  attachments: {
    name: string;
    type: string;
    attachmentId: string;
  }[];
}